+++ app/src/main/java/ru/surf/learn2invest/network_components/util/CoinRetrofitClient.kt
@@ -21,5 +21,5 @@ object CoinRetrofitClient {
         .build()
 }
 
-fun Double.round() =
+fun Float.round() =
     "\\S*\\.[0]*[0-9]{2}".toRegex().find(this.toBigDecimal().toPlainString())?.value
